requirements.txt
pyqrcode
pypng
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv~=0.14.0
asyncpg~=0.21.0
blockcypher~=1.0.90
requests
python-dateutil~=2.7.3
